+++ app/src/main/java/com/example/videomarker/data/entities/Data.java
@@ -1,4 +1,4 @@
-package com.example.videomarker.data;
+package com.example.videomarker.data.entities;
 
 public class Data {
